+++ Assets/Animation/Player/Fall.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 41446f207e360b24684af6950eaf1ecb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
