+++ Assets/Animation/Player/Fall/Fall.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 46541a91990cdbd459df76a587a5f47a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
