+++ Assets/Scripts/MAP/InterActiveItem/AutoRotate_Item.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e66f2d9c07948a14295aa7d3aeb6f4e9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
