--- Assets/Scripts/MAP/InterActiveItem/FloatBlock/AutoMove_Item.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3a3d4156895139c4ab9a823c1939f12a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
